# FIX Class manifest is not caching enums