--- a/src/Core/Manifest/ClassManifest.php
+++ b/src/Core/Manifest/ClassManifest.php
@@ -597,7 +597,7 @@ class ClassManifest
             $classes = $data['classes'];
             $interfaces = $data['interfaces'];
             $traits = $data['traits'];
-            $enums = $data['enums'];
+            $enums = $data['enums'] ?? [];
         } else {
             $changed = true;
             // Build from php file parser
@@ -774,7 +774,7 @@ class ClassManifest
         if (!$data || !is_array($data)) {
             return false;
         }
-        foreach (['classes', 'interfaces', 'traits', 'enums'] as $key) {
+        foreach (['classes', 'interfaces', 'traits'] as $key) {
             // Must be set
             if (!isset($data[$key])) {
                 return false;