@lekoala 2025-01-07
Actually now i see what enums are not there

https://github.com/silverstripe/silverstripe-framework/blob/f40b1a4ac061065fb38e19603b57b808b104061f/src/Core/Manifest/ClassManifest.php#L694-L701