@lekoala 2025-01-07
For reference, it's due to this PR
https://github.com/silverstripe/silverstripe-framework/commit/d3f104382d6eaa38caca03f6df34c2022dd02e80

which seems to have forgotten to store the enums in the cache

@kinglozzer did I miss something here?