@kinglozzer 2025-01-07
I can’t remember (or think of) any reason why they _shouldn’t_ be included in the cache, so it’s probably just an oversight in the original PR