@lekoala 2025-01-07
an additionnal improvement would be to not store empty arrays and simply consider missing keys as empty arrays.
this should reduce the cache size by a large margin, because it's currently storing a lot of empty traits => [] and enums => []