@lekoala 2025-01-10
since i made benchmarks for my other PR, i took the opportunity to benchmark this as well

without the fix

manifest (no cache folder)
Time: 8s
Peak memory: 16.000Mb
Memory: 16.00Mb

manifest (regenerate cache)
Time: 7s
Peak memory: 16.000Mb
Memory: 16.00Mb

with the fix

manifest (no cache folder)
Time: 9s
Peak memory: 16.000Mb
Memory: 16.00Mb

manifest (regenerate cache)
Time: 640ms
Peak memory: 14.000Mb
Memory: 14.00Mb

The missing cache key makes the cache regeneration take 7s instead of 640ms...